% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME_deconvolution.R
\name{abis}
\alias{abis}
\title{abis deconvolution}
\usage{
abis(sig_matrix, SE)
}
\arguments{
\item{sig_matrix}{gene expression matrix from isolated cells.}

\item{SE}{an SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}
}
\description{
use abis to predict TME
}
