% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{bayesmodel}
\alias{bayesmodel}
\title{Naive Bayes model for predicting ICI response.}
\format{
A model
}
\source{
Calculating
}
\description{
Built Naive Bayes model for predicting ICI response.
}
