% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bind_meta}
\alias{bind_meta}
\title{perform naive bayes prediction model.}
\usage{
bind_meta(SE, isList)
}
\arguments{
\item{SE}{an SummarizedExperiment object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{isList}{whether SE is list}
}
\description{
Generate a naive bayes model.
}
