% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bind_mtr}
\alias{bind_mtr}
\title{perform naive bayes prediction model.}
\usage{
bind_mtr(SE, isList)
}
\arguments{
\item{SE}{a SummarizedExperiment(SE) object or a list consists of multiple SE objects. The colData of the SE object(s) must contain treatment information named Treatment.}

\item{isList}{whether SE is list}
}
\description{
Generate a naive bayes model.
}
