% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_BE.R
\name{browse_BE}
\alias{browse_BE}
\title{Perform PCA to visualize batch effect}
\usage{
browse_BE(SE)
}
\arguments{
\item{SE}{a SummarizedExperiment(SE) object or a list consists of multiple SE objects. The colData of the SE object(s) must contain treatment information named Treatment.}
}
\description{
perform PCA to visualize batch effect.
}
