% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_biomk.R
\name{browse_biomk}
\alias{browse_biomk}
\title{dd}
\usage{
browse_biomk(SE, feature, colors = NULL, PT_drop = FALSE)
}
\arguments{
\item{SE}{dd}

\item{feature}{dd}

\item{colors}{description}

\item{PT_drop}{description}
}
\description{
dd
}
