% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Model.R
\name{build_CC_model}
\alias{build_CC_model}
\title{Build cancerclass prediction model for immunotherapy response}
\usage{
build_CC_model(SE, Signature, rmBE = TRUE, response_NR = TRUE)
}
\arguments{
\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{Signature}{an gene set you interested in}

\item{rmBE}{whether remove batch effect between different data set using internal Combat method}

\item{response_NR}{If TRUE, only use R or NR to represent Immunotherapy response of patients.}
}
\description{
Generate a cancerclass model.
}
