% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_CellType_Ref.R
\name{build_CellType_Ref}
\alias{build_CellType_Ref}
\title{Generate a custom reference matrix}
\usage{
build_CellType_Ref(
  Seurat_obj,
  logfc.threshold = 0.1,
  min.pct = 0.1,
  only.pos = TRUE
)
}
\arguments{
\item{Seurat_obj}{a Seurat object in which samples are clustered into different cell types.}

\item{logfc.threshold}{limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is 0.1 Increasing logfc.threshold speeds up the function, but can miss weaker signals. If the slot parameter is "scale.data" no filtering is performed.}

\item{min.pct}{only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations. Meant to speed up the function by not testing genes that are very infrequently expressed. Default is 0.01.}

\item{only.pos}{only return positive markers (FALSE by default).}
}
\description{
Generate a custom reference matrix from single-cell RNA sequencing data for TME deconvolution.
}
