% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Model.R
\name{build_Model.matrix}
\alias{build_Model.matrix}
\title{Build machine learning prediction model for immunotherapy response}
\usage{
build_Model.matrix(mtr, meta, Model, response_NR = TRUE, ...)
}
\arguments{
\item{mtr}{the dataset you wish to use to build your model. A SummarizedExperiment (SE) object, which can be either a single SE object or a list of SE objects. Note that for each SE object, the colData must contain treatment information under the column name Treatment.}

\item{meta}{refers to the specific set of genes you wish to use for model construction.}

\item{Model}{represents the type of model you want to build. You have several options to choose from: "NB" for Naive Bayes, "SVM" for Support Vector Machine, "RF" for Random Forest, "CC" for Cancerclass, "ADB" for Adaboost, "LGB" for Logitboost, and "LGT" for Logistics.}

\item{response_NR}{If TRUE, only use R or NR to represent Immunotherapy response of patients.}

\item{...}{the arguments}
}
\description{
Generate immunotherapy prognosis prediction model.
}
