% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{build_Model}
\alias{build_Model}
\title{Build naive bayes prediction model for immunotherapy response}
\usage{
build_Model(Model, SE, Signature, rmBE = FALSE, response_NR = TRUE)
}
\arguments{
\item{Model}{the type of model (NB-Naive bayes, SVM-Support Vector machine, RF-Random Forest, CC-Cancerclass, ADB-Adaboost, LGB-Logitboost, LGT-Logistics)}

\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{Signature}{an gene set you interested in}

\item{rmBE}{whether remove batch effect between different data set using internal Combat method}

\item{response_NR}{If TRUE, only use R or NR to represent Immunotherapy response of patients.}
}
\description{
Generate a naive bayes model.
}
