% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Model.R
\name{build_NB_model}
\alias{build_NB_model}
\title{Build naive bayes prediction model for immunotherapy response}
\usage{
build_NB_model(
  SE,
  Signature,
  rmBE = FALSE,
  response_NR = TRUE,
  laplace = 1,
  PT_drop,
  ...
)
}
\arguments{
\item{SE}{a SummarizedExperiment(SE) object or a list consists of multiple SE objects. The colData of the SE object(s) must contain treatment information named Treatment.}

\item{Signature}{an gene set you interested in}

\item{rmBE}{whether remove batch effect between different data sets using internal Combat method.}

\item{response_NR}{If TRUE, classify patients with CR, MR, PR as Responders (R), and those with PD, SD, NR as Non-Responders(NR).}

\item{laplace}{positive double controlling Laplace smoothing. The default (0) disables Laplace smoothing.}

\item{PT_drop}{If TRUE, only Untreated patient will be use for model training.}

\item{...}{the arguments}
}
\description{
Generate a naive bayes model.
}
