% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{build_RF_model}
\alias{build_RF_model}
\title{perform ramdom forest prediction model.}
\usage{
build_RF_model(SE, Signature, rmBE = TRUE)
}
\arguments{
\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{Signature}{an gene set you interested in}

\item{rmBE}{whether remove batch effect between different data set using internal Combat method}
}
\description{
Generate a random forest model.
}
