% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_TME.R
\name{Ciber}
\alias{Ciber}
\title{Cibersort functions}
\usage{
Ciber(sig_matrix, mix_matrix, perm = 0, QN = TRUE)
}
\arguments{
\item{sig_matrix}{file path to gene expression from isolated cells}

\item{mix_matrix}{heterogenous mixed expression}

\item{perm}{Number of permutations}

\item{QN}{Perform quantile normalization or not (TRUE/FALSE)}
}
\description{
Cibersort functions which perform deconvolution to bulk RNA-seq data. And return the cell fraction.
}
