% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_biomk.R
\name{compare_biomk}
\alias{compare_biomk}
\title{Compare the AUC of signature between datasets}
\usage{
compare_biomk(
  SE = NULL,
  Signature = NULL,
  method = "Average_mean",
  PT_drop = TRUE,
  show.val = TRUE,
  val.size = 2
)
}
\arguments{
\item{SE}{a SummarizedExperiment object for which you want to calculate the Signature Score.}

\item{Signature}{a genes vector represents user-defined signature for Immunotherapy response. If NULL, the function will only calculate 23 built-in signatures in tigeR.}

\item{method}{the method for calculating gene set scores which has several options: Average_mean, Weighted_mean, or GSVA. The method can be set to NULL if the length of the parameter geneSet is 1. This means that if you are working with only one gene, the specific calculation method may not be applicable or necessary.}

\item{PT_drop}{If TRUE, only Untreated patient will be use for model training.}

\item{show.val}{If TRUE, the value will be show in the heatplot.}

\item{val.size}{an integer represents the size of AUC value.}
}
\description{
generate a heatmap of signature auc of datasets.
}
