% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Core}
\alias{Core}
\title{count geneset score by different method}
\usage{
Core(exp_mtr, geneSet, method)
}
\arguments{
\item{exp_mtr}{an expression matrix.}

\item{geneSet}{the gene or geneset which you wanted to investigate.}

\item{method}{the method for calculating gene set scores. Can be NULL if the length of parameter gene is 1.}
}
\description{
wait to write
}
