% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME_deconvolution.R
\name{CoreAlg}
\alias{CoreAlg}
\title{core algorithm}
\usage{
CoreAlg(X, Y)
}
\arguments{
\item{X}{cell-specific gene expression}

\item{Y}{mixed expression per sample}
}
\description{
core algorithm of CIBERSORT
}
