% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardization.R
\name{data_standardization}
\alias{data_standardization}
\title{Build machine learning prediction model for immunotherapy response}
\usage{
data_standardization(SE, type)
}
\arguments{
\item{SE}{the dataset you wish to use to build your model. A SummarizedExperiment (SE) object, which can be either a single SE object or a list of SE objects. Note that for each SE object, the colData must contain treatment information under the column name Treatment.}

\item{type}{can be c(1,2,3,4)}
}
\description{
Generate immunotherapy prognosis prediction model.
}
