% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataloader.R
\name{Dataloader}
\alias{Dataloader}
\title{Process data before running machine learning algorithm}
\usage{
Dataloader(pick = NULL)
}
\arguments{
\item{pick}{the number of dataset you want to download. If NULL, return the dataset list.}
}
\description{
Process data before running machine learning algorithm
}
