% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocess_SE.R
\name{dataPreprocess_SE}
\alias{dataPreprocess_SE}
\title{Prepare expression matrix for down string analysis}
\usage{
dataPreprocess_SE(SE, Signature = NULL, turn2HL = TRUE)
}
\arguments{
\item{SE}{d}

\item{Signature}{The aiming gene set(only Gene SYMBOL allowed).}

\item{turn2HL}{If TRUE, the expression value of a gene is divided to "HIGH" or "LOW" based on its median expression.}
}
\description{
dataPreprocess will remove missing genes. Then returns the sub-matrix of the genes whose SYMBOLs are in the signature.
}
