% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunotherapy_Response.R
\name{DEA_Response}
\alias{DEA_Response}
\title{Calculating differential expression score between responder and non_responder.}
\usage{
DEA_Response(SE, gene)
}
\arguments{
\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information names response.}

\item{gene}{The gene which you wanted.}
}
\description{
Differential expression score was alculated using the following formula: âˆ’ğ‘†ğ¼ğºğ‘(ğ‘™ğ‘œğ‘”2(ğ¹ğ¶)) Ã— ğ‘™ğ‘œğ‘”10(ğ‘ƒ), where ğ¹ğ¶ represents the fold change and ğ‘ƒ represents the P value derived from the Wilcoxon rank-sum test
}
