% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_TME.R
\name{deconv_TME}
\alias{deconv_TME}
\title{Perform tumor microenvironment deconvolution}
\usage{
deconv_TME(SE, method, ...)
}
\arguments{
\item{SE}{a SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}

\item{method}{the TME analysis method you want to apply for.}

\item{...}{other argument in the 10 TME analysis functions.}
}
\description{
Perform tumor microenvironment deconvolution through 10 open-source algorithms.
}
