% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_biomk.R
\name{diff_biomk}
\alias{diff_biomk}
\title{Visualize differential analysis result (Responder vs NonResponder or Pre-Treatment vs Post-Treatment).}
\usage{
diff_biomk(
  SE,
  gene,
  type,
  method = "Average_mean",
  PT_drop = TRUE,
  p.pos = c(0.2, 0.7),
  p.round = 2,
  log_sc = TRUE,
  textcol = "black",
  panelcol = "black"
)
}
\arguments{
\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{gene}{the Gene or Gene set you are interested in.}

\item{type}{the comparison group you want to choose, 'Treatment' (Pre-Treatment vs Post-Treatment) or 'Response' (Responder vs Non-Responder ).}

\item{method}{the method for calculating gene set scores which has several options: Average_mean, Weighted_mean, or GSVA. The method can be set to NULL if the length of the parameter geneSet is 1. This means that if you are working with only one gene, the specific calculation method may not be applicable or necessary.}

\item{PT_drop}{If TRUE, only Untreated patient will be use for model training.}

\item{p.pos}{the position of the P value. When p.pos=c(0,0), the legend will be placed in the bottom left corner of the plot.}

\item{p.round}{the decimal places you want to keep for p value}

\item{log_sc}{if TRUE, log(value + 1)}

\item{textcol}{the color of the text in the plot.}

\item{panelcol}{the color of the panel border and ticks in the plot.}
}
\value{
\describe{
  Return a bar plot visualizing the differential analysis (Responder vs NonResponder or Pre-Treatment vs Post-Treatment)}
}
\description{
Visualize differential analysis result (Responder vs NonResponder or Pre-Treatment vs Post-Treatment).
}
\examples{
diff_biomk(SE=MEL_GSE78220,gene='CD274',type='Treatment') +
ggtitle("Pre-Treatment vs Post-Treatment") +
 theme(plot.title = element_text(hjust = 0.5))
}
