% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{diff_gene}
\alias{diff_gene}
\title{differential gene}
\usage{
diff_gene(SE)
}
\arguments{
\item{SE}{a SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}
}
\description{
Return differential expression gene between Responder and Non-Responder
}
