% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_TME.R
\name{diff_TME}
\alias{diff_TME}
\title{Visualize difference of cell fraction}
\usage{
diff_TME(
  SE,
  feature = NULL,
  style = "elegant",
  group_color = c("#5f96e8CC", "#ee822fCC")
)
}
\arguments{
\item{SE}{a SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{feature}{the feature or features you are interested in.}

\item{style}{the plotting style. Choose from c("elegant", "raw").}

\item{group_color}{a vector of colors with length 2.
The first color corresponds to Non-Responders and
the second color corresponds to Responders.}
}
\description{
Visualize difference of cell fraction between Responder and Non-Responder or Pre-Treatment and Post-Treatment patients.
}
