% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_TME.R
\name{EPIC}
\alias{EPIC}
\title{epic deconvolution}
\usage{
EPIC(SE, sig_matrix = NULL, ...)
}
\arguments{
\item{SE}{a SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}

\item{sig_matrix}{description}

\item{...}{the arguments}
}
\description{
use epic to predict TME
}
