% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_TME.R
\name{ESTIMATE}
\alias{ESTIMATE}
\title{ESTIMATE deconvolution}
\usage{
ESTIMATE(SE)
}
\arguments{
\item{SE}{a SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}
}
\description{
use ESTIMATE to predict TME
}
