% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_label}
\alias{extract_label}
\title{Binding response data from data folder in tigeR together}
\usage{
extract_label(datasetNames)
}
\arguments{
\item{datasetNames}{the name of data set or data sets you want to use.}
}
\description{
Extract response data in particular data set or data sets from the data folder in tigeR. If there are more than one data set, this function will return an vector which contains the response data of every data sets.
}
