% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_pie.R
\name{fraction_pie}
\alias{fraction_pie}
\title{draw integrate fraction pie}
\usage{
fraction_pie(mtr, feature)
}
\arguments{
\item{mtr}{cell fraction matrix}

\item{feature}{the factor of color levels}
}
\description{
draw integrate fraction pie
}
