% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{geneCorr}
\alias{geneCorr}
\title{Calculate the correlations between user defined gene or gene set and TCGA cancerous samples.}
\usage{
geneCorr(gene = "CD274")
}
\arguments{
\item{gene}{is the Gene or Gene set you are interested in.}
}
\description{
Generating an Heatmap showing the pearson correlation coefficient.
}
