% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_Model.R
\name{generate_result}
\alias{generate_result}
\title{Test prediction model for immunotherapy response}
\usage{
generate_result(ROC)
}
\arguments{
\item{ROC}{the roc object}
}
\description{
Test a naive bayes model.
}
