% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R, R/plot.R
\name{geneSurv}
\alias{geneSurv}
\title{Calculate the association between gene expression and overall survival in the immunotherapy data.}
\usage{
geneSurv(gene = "CD274", type = "cox")

geneSurv(gene = "CD274", type = "cox")
}
\arguments{
\item{gene}{is the Gene or Gene set you are interested in.}

\item{type}{is the analysis you want to perform.'cox' or 'auc'}
}
\description{
The association between gene expression and overall survival in the immunotherapy data was calculated using univariate Cox regression analysis.

The association between gene expression and overall survival in the immunotherapy data was calculated using univariate Cox regression analysis.
}
