% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Gini}
\alias{Gini}
\title{Ranking features in vector with Gini coefficient}
\usage{
Gini(vec, label)
}
\arguments{
\item{vec}{An vector. Usually is one row of an matrix.}

\item{label}{The classification label of matrix.}
}
\description{
Ranking features in vector with Gini coefficient
}
