% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Gini_rank}
\alias{Gini_rank}
\title{Ranking features in matrix with Gini index}
\usage{
Gini_rank(mtr, label, ascending = TRUE)
}
\arguments{
\item{mtr}{An gene expression matrix which rownames are gene symbols and colnames are sample ID.}

\item{label}{The classification labels of the samples.}

\item{ascending}{If ascending = TRUE, the result will be display in ascending order.}
}
\description{
By calculating the Gini index of different genes, you can get an overview of the classification efficiency of different genes.
}
