% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunotherapy_Response.R
\name{Immunotherapy_Response}
\alias{Immunotherapy_Response}
\title{Perform differential expression analysis and survival analysis.}
\usage{
Immunotherapy_Response(SE, geneSet = NULL, method)
}
\arguments{
\item{SE}{a SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain treatment information names Treatment.}

\item{geneSet}{The geneSet which you wanted.}

\item{method}{the method for calculating gene set scores. Can be NULL if the length of parameter gene is 1.}
}
\description{
Perform differential expression analysis and survival analysis in certain gene and return the result.
}
