% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_analysis.R
\name{integrate_analysis}
\alias{integrate_analysis}
\title{Perform differential expression analysis and survival analysis.}
\usage{
integrate_analysis(SE, geneSet = NULL, method = NULL, PT_drop = TRUE)
}
\arguments{
\item{SE}{a SummarizedExperiment(SE) object or a list consists of multiple SE objects. The colData of the SE object(s) must contain treatment information named Treatment.}

\item{geneSet}{the geneSet which you wanted.}

\item{method}{the method for calculating gene set scores which has several options: "Average_mean", "Weighted_mean", or "GSVA". The method can be set to NULL if the length of the parameter geneSet is 1. This means that if you are working with only one gene, the specific calculation method may not be applicable or necessary.}

\item{PT_drop}{if TRUE, only Untreated patient will be use for model training.}
}
\description{
Perform differential expression analysis and survival analysis for a gene or gene set.
}
\details{
\bold{Algorithm}

The score of differential expression analysis is calculated by following formula:

\eqn{ -\text{SIGN}(\log_2(FC)) \times \log_{10}(p) }

where \emph{FC} represents the fold change and \emph{p} represents the P value derived from the Wilcoxon rank-sum test.

The score of survival analysis is calculated by the following formula:

\eqn{ -\text{SIGN}(\log_2(HR)) \times \log_{10}(p) }

where \emph{HR} represents the hazard ratio and \emph{p} represents the P value derived from univariate Cox regression analysis.
}
\examples{
# Single-Gene Analysis
integrate_analysis(SE=MEL_GSE78220, geneSet="CD274")
}
