% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME_deconvolution.R
\name{IPS}
\alias{IPS}
\title{xCell deconvolution}
\usage{
IPS(SE, project = NULL, plot = FALSE)
}
\arguments{
\item{SE}{an SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}

\item{signatures}{a GMT object of signatures.}

\item{genes}{list of genes to use in the analysis.}

\item{spill}{the Spillover object for adjusting the scores.}

\item{rnaseq}{if true than use RNAseq spillover and calibration paramters, else use array parameters.}

\item{file.name}{string for the file name for saving the scores. Default is NULL.}

\item{scale}{if TRUE, uses scaling to trnasform scores using fit.vals}

\item{alpha}{a value to override the spillover alpha parameter. Deafult = 0.5}

\item{save.raw}{TRUE to save a raw}

\item{parallel.sz}{integer for the number of threads to use. Default is 4.}

\item{parallel.type}{Type of cluster architecture when using snow. 'SOCK' or 'FORK'. Fork is faster, but is not supported in windows.}

\item{cell.types.use}{a character list of the cell types to use in the analysis. If NULL runs xCell with all cell types.}
}
\description{
use xCell to predict TME
}
