% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signatures_grading.R
\name{IRS_grading}
\alias{IRS_grading}
\title{Giving immunotherapy prognosis using IRS.}
\usage{
IRS_grading(exp_mtr)
}
\arguments{
\item{exp_mtr}{An expression matrix which rownames are gene symbol and colnames are sample ID.}
}
\description{
The function will return a vector calculated using IRS.
}
