% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.NA_vec}
\alias{is.NA_vec}
\title{judge whether all the elements in vector V are NA}
\usage{
is.NA_vec(V)
}
\arguments{
\item{V}{must be a vector.}
}
\description{
judge whether all the elements in vector V are NA
}
