\name{LCS}
\alias{LCS}
\title{ Longest common sequence}
\description{
   Faster implementation of the function provided by the qualV-package
}
\usage{
LCS(a, b)
}
\arguments{
  \item{a}{ first vector. }
  \item{b}{ second vector. }
}
\seealso{\code{\link[qualV]{LCS}}}
\author{ Dominik Reusser }
\keyword{ utilities }

