% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitBoost.R
\name{LogitBoost}
\alias{LogitBoost}
\title{Test prediction model for immunotherapy response}
\usage{
LogitBoost(xlearn, ylearn, nIter)
}
\arguments{
\item{xlearn}{model generated by build_Model() function}

\item{ylearn}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{nIter}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}
}
\description{
Test a naive bayes model.
}
