% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_analysis.R
\name{matrix_cox}
\alias{matrix_cox}
\title{cox regression}
\usage{
matrix_cox(V, meta)
}
\arguments{
\item{V}{the expression vector of a gene}

\item{meta}{the meta information.}
}
\description{
wait to write
}
