% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{max_min_normalization}
\alias{max_min_normalization}
\title{Max_Min normalization.}
\usage{
max_min_normalization(exp_mtr)
}
\arguments{
\item{exp_mtr}{an matrix which rows represent genes columns represent samples.}
}
\description{
(x - min(x))/(max(x) - min(x))
}
