% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME_deconvolution.R
\name{MCPcounter.estimate}
\alias{MCPcounter.estimate}
\title{ESTIMATE deconvolution}
\usage{
MCPcounter.estimate(expression, featuresType, probesets, genes)
}
\arguments{
\item{expression}{matrix or data.frame with features in rows and samples in columns}

\item{featuresType}{type of identifiers for expression features. Defaults to "affy133P2_probesets" for Affymetrix Human Genome 133 Plus 2.0 probesets. Other options are "HUGO_symbols" (Official gene symbols), "ENTREZ_ID" (Entrez Gene ID) or "ENSEMBL_ID" (ENSEMBL Gene ID)}

\item{probesets}{probs}

\item{genes}{genes}
}
\description{
use ESTIMATE to predict TME
}
