% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME_deconvolution.R
\name{MCPCounter}
\alias{MCPCounter}
\title{MCPCounter deconvolution}
\usage{
MCPCounter(SE, featuresType = "HUGO_symbols", ...)
}
\arguments{
\item{SE}{an SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}

\item{featuresType}{type of identifiers for expression features. Defaults to "affy133P2_probesets" for Affymetrix Human Genome 133 Plus 2.0 probesets. Other options are "HUGO_symbols" (Official gene symbols), "ENTREZ_ID" (Entrez Gene ID) or "ENSEMBL_ID" (ENSEMBL Gene ID)}

\item{...}{other parameter}
}
\description{
use MCPCounter to predict TME
}
