% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_Engineering.R
\name{merge_SE}
\alias{merge_SE}
\title{Merge SummarizedExperiment object}
\usage{
merge_SE(SE)
}
\arguments{
\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}
}
\description{
Merge SummarizedExperiment object
}
