% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signatures_grading.R
\name{MSKCC_grading}
\alias{MSKCC_grading}
\title{Giving immunotherapy prognosis using MSKCC.}
\usage{
MSKCC_grading(exp_mtr)
}
\arguments{
\item{exp_mtr}{An expression matrix which rownames are gene symbol and colnames are sample ID.}
}
\description{
The function will return a vector calculated using MSKCC.
}
