% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signatures_grading.R
\name{PanFTBRs_grading}
\alias{PanFTBRs_grading}
\title{Giving immunotherapy prognosis using Pan_F_TBRs.}
\usage{
PanFTBRs_grading(exp_mtr)
}
\arguments{
\item{exp_mtr}{An expression matrix which rownames are gene symbol and colnames are sample ID.}
}
\description{
The function will return a vector calculated using Pan_F_TBRs.
}
