% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT.R
\name{plt_CIBERSORT}
\alias{plt_CIBERSORT}
\title{Process data before running machine learning algorithm}
\usage{
plt_CIBERSORT(sig_matrix, SE, perm = 0, QN = TRUE)
}
\arguments{
\item{sig_matrix}{file path to gene expression from isolated cells}

\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{perm}{Number of permutations}

\item{QN}{Perform quantile normalization or not (TRUE/FALSE)}
}
\description{
Process data before running machine learning algorithm
}
