% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plt_diff}
\alias{plt_diff}
\title{plot differential result(Responder vs Non-Responder or Pre-Treatment vs Post-Treatment)}
\usage{
plt_diff(gene = "CD274", SE, type)
}
\arguments{
\item{gene}{is the Gene or Gene set you are interested in.}

\item{SE}{SE an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{type}{'Treatment' or 'Response'.the type of analysis you want to perform(Responder vs Non-Responder or Pre-Treatment vs Post-Treatment)}
}
\description{
plot differential result(Responder vs Non-Responder or Pre-Treatment vs Post-Treatment)
}
