% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plt_roc}
\alias{plt_roc}
\title{Prepare data for plot}
\usage{
plt_roc(ROC, auc.pos, auc.round, textcol)
}
\arguments{
\item{ROC}{the ROC object}

\item{auc.pos}{the position of the AUC value}

\item{auc.round}{the decimal places you want to keep for auc value}

\item{textcol}{the color of the text in the plot}
}
\description{
Preparing data for ploting.
}
