% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plt_RvsNR}
\alias{plt_RvsNR}
\title{plot differential result(responder vs nonresponder)}
\usage{
plt_RvsNR(gene = "CD274", SE)
}
\arguments{
\item{gene}{is the Gene or Gene set you are interested in.}

\item{SE}{SE an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}
}
\description{
The association between gene expression and overall survival in the immunotherapy data was calculated using univariate Cox regression analysis.
}
