% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_prediction.R
\name{pred_response}
\alias{pred_response}
\title{Predict immunotherapy response for individual patient}
\usage{
pred_response(
  SE = NULL,
  exp_mtr = NULL,
  meta = NULL,
  threshold = 0.8,
  positive = "R",
  Signature = NULL,
  method = "Average_mean",
  PT_drop = TRUE,
  show.val = TRUE,
  sort_by = "rankscore",
  group_by = "rankscore",
  show.Observed = TRUE,
  text_col = "black",
  rankscore = TRUE,
  ZS = TRUE
)
}
\arguments{
\item{SE}{a SummarizedExperiment object for which you want to calculate the signature score.}

\item{exp_mtr}{an expression matrix for which you want to calculate the Signature Score.}

\item{meta}{meta data of samples}

\item{threshold}{the threshold for custom signature to discretize patients into R and N group.}

\item{positive}{c("N","R"), the positive case.}

\item{Signature}{a gene vector represents user-defined signature for immunotherapy response. If NULL, the function will only calculate 23 built-in signatures in tigeR.}

\item{method}{the method for calculating gene set scores which has several options: "Average_mean", "Weighted_mean", or "GSVA". The method can be set to NULL if the length of the parameter geneSet is 1. This means that if you are working with only one gene, the specific calculation method may not be applicable or necessary.}

\item{PT_drop}{if TRUE, only Untreated patient will be use for model training.}

\item{show.val}{if TRUE, the value will be show in the heatplot.}

\item{sort_by}{the signature for the sample sorting in the heatmap.}

\item{group_by}{the signature score used for response classification.}

\item{show.Observed}{if TRUE, show the observed cases in heatmap.}

\item{text_col}{the color of the text in heatmap.}

\item{rankscore}{if TRUE show the Robust Rank Aggregation score in heatmap.}

\item{ZS}{if TRUE, conduct Z-score scaling on each signature score.}
}
\description{
Predict immunotherapy response and generate heatmap of signatures.
}
