% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{predict_Response}
\alias{predict_Response}
\title{perform naive bayes prediction model.}
\usage{
predict_Response(exp, response, Signature, rmBE = TRUE)
}
\arguments{
\item{exp}{a matrix or list consist of matrix. The rownames of every matrix must be GENE SYMBOL. If it is a list, the row num and row names must be the same.}

\item{response}{an character vector corresponding to samples' clinical status (respond or not,survival or dead etc.)}

\item{Signature}{an gene set you interested in}

\item{rmBE}{whether remove batch effect between different data set using internal Combat method}
}
\description{
Generate a naive bayes model.
}
