% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signatures_grading.R
\name{predict_Signature}
\alias{predict_Signature}
\title{Immunotherapy prognosis .}
\usage{
predict_Signature(exp_mtr)
}
\arguments{
\item{exp_mtr}{An expression matrix which rownames are gene symbol and colnames are sample ID.}
}
\description{
The function will return a vector calculated using T cell–inflamed GEP score.
}
