% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signatures_grading.R
\name{PRS_grading}
\alias{PRS_grading}
\title{Giving immunotherapy prognosis using PRS Signature.}
\usage{
PRS_grading(exp_mtr)
}
\arguments{
\item{exp_mtr}{An expression matrix which rownames are gene symbol and colnames are sample ID.}
}
\description{
The function will return a vector calculated using PRS Signature.
}
