% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME_deconvolution.R
\name{quanTIseq}
\alias{quanTIseq}
\title{quanTIseq deconvolution}
\usage{
quanTIseq(SE, tumor = TRUE, arrays = FALSE, scale_mrna = FALSE, ...)
}
\arguments{
\item{SE}{a SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}

\item{tumor}{whether the input is tumor sample}

\item{arrays}{whether you data is array data}

\item{scale_mrna}{whether perform scaling}

\item{...}{other parameter}
}
\description{
use quanTIseq to predict TME
}
