% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_Reference.R
\name{refine_Reference}
\alias{refine_Reference}
\title{Refine a custom reference matrix}
\usage{
refine_Reference(
  Seurat_obj,
  n_ref_genes = 50,
  logfc.threshold = 0.15,
  min.pct = 0.1,
  only.pos = TRUE
)
}
\arguments{
\item{Seurat_obj}{des}

\item{n_ref_genes}{description}

\item{logfc.threshold}{description}

\item{min.pct}{des}

\item{only.pos}{description}
}
\description{
refine a Custom reference matrix for TME deconvolution from single-cell sequencing data.
}
