% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_BE.R
\name{remove_BE}
\alias{remove_BE}
\title{Remove batch effect}
\usage{
remove_BE(SE, method)
}
\arguments{
\item{SE}{a SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{method}{batch effect correction methods.}
}
\description{
remove batch effect.
}
