% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{response_filter}
\alias{response_filter}
\title{Filting missing response value.}
\usage{
response_filter(response)
}
\arguments{
\item{response}{a vector which contains response information.}
}
\description{
return the index of NE or UNK in response vector.
}
