% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{response_standardize}
\alias{response_standardize}
\title{standardization of response labels}
\usage{
response_standardize(V)
}
\arguments{
\item{V}{an vector}
}
\description{
turn label to 'R' or 'NR'
}
