% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_biomk.R
\name{roc_biomk}
\alias{roc_biomk}
\title{Assessing the performance of signature on predicting immunotherapy response}
\usage{
roc_biomk(
  SE,
  Signature,
  method = NULL,
  rmBE = FALSE,
  response_NR = TRUE,
  PT_drop = TRUE,
  auc.pos = c(0.3, 0.42),
  auc.round = 3,
  textcol = "black",
  panelcol = "black"
)
}
\arguments{
\item{SE}{the dataset you use to assess your signature. A SummarizedExperiment (SE) object, which can be either a single SE object or a list of SE objects. Note that for each SE object, the colData must contain treatment information under the column named Treatment.}

\item{Signature}{a gene vector represents the built-in or user-defined signature.}

\item{method}{the method for calculating gene set scores which has several options: "Average_mean", "Weighted_mean", or "GSVA". The method can be set to NULL if the length of the parameter geneSet is 1. This means that if you are working with only one gene, the specific calculation method may not be applicable or necessary.}

\item{rmBE}{whether remove batch effect between different data set using internal Combat method}

\item{response_NR}{If TRUE, classify patients with CR, MR, PR as Responders (R), and those with PD, SD, NR as Non-Responders(NR).}

\item{PT_drop}{If TRUE, only untreated patient will be use for model training.}

\item{auc.pos}{the position of the AUC value}

\item{auc.round}{the decimal places you want to keep for auc value}

\item{textcol}{the color of the text in the plot.}

\item{panelcol}{the color of the panel border and ticks in the plot.}
}
\value{
\describe{
  Return a list contain the following elements which including:
    \item{\code{ROC}}{a receiver operating characteristic (ROC) object.}
    \item{\code{figure}}{a roc curve to show the predictive performance.}
  }
}
\description{
generate a Receiver Operating Characteristic (ROC) object and a curve to assess the predictive performance.
}
\examples{
sig_roc <-
roc_biomk(MEL_GSE93157,
         Weighted_mean_Sigs$Tcell_inflamed_GEP,
         method = "Weighted_mean",
         rmBE=TRUE,
         response_NR=TRUE)
sig_roc
}
