% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_biomk_SE.R
\name{score_biomk_SE}
\alias{score_biomk_SE}
\title{Calculating Signature score of existing immunotherapy response Signature.}
\usage{
score_biomk_SE(
  SE = NULL,
  exp_mtr = NULL,
  meta = NULL,
  Signature = NULL,
  method = "Average_mean",
  PT_drop = TRUE
)
}
\arguments{
\item{SE}{a SummarizedExperiment object for which you want to calculate the Signature Score.}

\item{exp_mtr}{an expression matrix for which you want to calculate the Signature Score.}

\item{meta}{meta data of samples}

\item{Signature}{a genes vector represents user-defined signature for Immunotherapy response. If NULL, the function will only calculate 23 built-in signatures in tigeR.}

\item{method}{the method for calculating gene set scores which has several options: "Average_mean", "Weighted_mean", or "GSVA". The method can be set to NULL if the length of the parameter geneSet is 1. This means that if you are working with only one gene, the specific calculation method may not be applicable or necessary.}

\item{PT_drop}{If TRUE, only Untreated patient will be use for model training.}
}
\description{
By employing the Signature_calculation() function, you can obtain a comprehensive signature score matrix for the 23 signatures in tigeR. In this matrix, the columns represent the signature scores, and the rows denote the sample names.
}
