% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_Engineering.R
\name{SE_Merging}
\alias{SE_Merging}
\title{Merge SummarizedExperiment object}
\usage{
SE_Merging(SE)
}
\arguments{
\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}
}
\description{
Merge SummarizedExperiment object
}
