% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature_assessment.R
\name{Signature_assessment}
\alias{Signature_assessment}
\title{Assessment of gene set}
\usage{
Signature_assessment(SE, Signature, rmBE, response_NR)
}
\arguments{
\item{SE}{the dataset you wish to use to test your Signature. A SummarizedExperiment (SE) object, which can be either a single SE object or a list of SE objects. Note that for each SE object, the colData must contain treatment information under the column name Treatment.}

\item{Signature}{The gene which you wanted.}

\item{rmBE}{whether remove batch effect between different data set using internal Combat method}

\item{response_NR}{If TRUE, only use R or NR to represent Immunotherapy response of patients.}
}
\description{
Perform differential expression analysis and survival analysis in certain gene and return the result.
}
