% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature_calculation.R
\name{Signature_calculation}
\alias{Signature_calculation}
\title{Calculating Signature score of existing immunotherapy prognosis Signature.}
\usage{
Signature_calculation(SE = NULL, exp_mtr = NULL)
}
\arguments{
\item{SE}{a SummarizedExperiment object for which you want to calculate the Signature Score.}

\item{exp_mtr}{an expression matrix for which you want to calculate the Signature Score.}
}
\description{
By employing the Signature_calculation() function, you can obtain a comprehensive signature score matrix for the 23 signatures in TigeR. In this matrix, the columns represent the signature scores, and the rows denote the sample names.
}
