% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_biomk.R
\name{surv_biomk}
\alias{surv_biomk}
\title{perform survival analysis}
\usage{
surv_biomk(
  SE,
  gene,
  method = "Average_mean",
  style = "elegant",
  conf.int = FALSE,
  val.pos = c(0.03, 0.2),
  lg = NULL,
  lg.pos = c(0.6, 0.9),
  lg.text = "precise",
  p.round = 4,
  PT_drop = TRUE,
  drop_zero = FALSE
)
}
\arguments{
\item{SE}{SE an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{gene}{is the Gene or Gene set you are interested in.}

\item{method}{the method for calculating gene set scores which has several options: Average_mean, Weighted_mean, or GSVA. The method can be set to NULL if the length of the parameter geneSet is 1. This means that if you are working with only one gene, the specific calculation method may not be applicable or necessary.}

\item{style}{the ploting style. c("raw,"elegant","brief")}

\item{conf.int}{logical value. If TRUE, plots confidence interval.}

\item{val.pos}{the position of annotation value.}

\item{lg}{dd}

\item{lg.pos}{the position of legend. When lg.pos=c(0,0), the legend will be placed at the leftdown of the plot.}

\item{lg.text}{c("precise","all"). If all, show the 0.95 CI of median.}

\item{p.round}{the decimal places you want to keep for p value}

\item{PT_drop}{If TRUE, only Untreated patient will be use for model training.}

\item{drop_zero}{If TRUE, 0 will not be consider when calculate median.}
}
\description{
calculates hazard ratios, confidence intervals and P value of cox-ph analysis as well as draw KM curve.
}
