% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_biomk.R
\name{surv_styling}
\alias{surv_styling}
\title{the style of survival plot}
\usage{
surv_styling(
  df,
  style,
  conf.int,
  gene,
  method,
  val.pos,
  lg,
  lg.pos,
  lg.text,
  p.round
)
}
\arguments{
\item{df}{the data source}

\item{style}{plotting style}

\item{conf.int}{logical value. If TRUE, plots confidence interval.}

\item{gene}{the gene you interested in.}

\item{method}{the method for calculating gene set scores which has several options: "Average_mean", "Weighted_mean", or "GSVA". The method can be set to NULL if the length of the parameter geneSet is 1. This means that if you are working with only one gene, the specific calculation method may not be applicable or necessary.}

\item{val.pos}{the position of annotation value.}

\item{lg}{dd}

\item{lg.pos}{the position of legend.}

\item{lg.text}{c("precise","all"). If all, show the 0.95 CI of median.}

\item{p.round}{the decimal places you want to keep for p value}
}
\description{
the style of survival plot
}
