% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunotherapy_Response.R
\name{survival_Score}
\alias{survival_Score}
\title{Calculating survival score of patients.}
\usage{
survival_Score(SE, geneSet, method)
}
\arguments{
\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain event time(names time), event(names status).}

\item{geneSet}{The geneSet which you wanted.}

\item{method}{the method for calculating gene set scores. Can be NULL if the length of parameter gene is 1.}
}
\description{
Survival score was calculated using the following formula: âˆ’ğ‘†ğ¼ğºğ‘(ğ‘™ğ‘œğ‘”2(ğ»ğ‘…)) Ã— ğ‘™ğ‘œğ‘”10(ğ‘ƒ), where ğ»ğ‘… represents the hazard ratio and ğ‘ƒ represents the P value derived from univariate Cox regression analysis.
}
