% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signatures_grading.R
\name{Tcell_inflamed_GEP_grading}
\alias{Tcell_inflamed_GEP_grading}
\title{Giving immunotherapy prognosis using T cell–inflamed GEP score.}
\usage{
Tcell_inflamed_GEP_grading(exp_mtr)
}
\arguments{
\item{exp_mtr}{An expression matrix which rownames are gene symbol and colnames are sample ID.}
}
\description{
The function will return a vector calculated using T cell–inflamed GEP score.
}
