% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_Model.R
\name{test_Model.default}
\alias{test_Model.default}
\title{Test prediction model for immunotherapy response}
\usage{
test_Model.default(Model, SE, PT_drop = TRUE)
}
\arguments{
\item{Model}{the model that you want to test, which is generated by the build_Model() function.}

\item{SE}{the dataset you wish to use to test your model. A SummarizedExperiment (SE) object, which can be either a single SE object or a list of SE objects. Note that for each SE object, the colData must contain treatment information under the column name Treatment.}

\item{PT_drop}{If TRUE, only Untreated patient will be use for model training.}
}
\description{
Generate a naive bayes model.
}
