% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_Model.R
\name{test_Model.matrix}
\alias{test_Model.matrix}
\title{Test prediction model for immunotherapy response}
\usage{
test_Model.matrix(mtr, meta, Model)
}
\arguments{
\item{mtr}{the result matrix of TME deconvolution.}

\item{meta}{meta data of samples}

\item{Model}{the model that you want to test, which is generated by the build_Model() function.}
}
\description{
Generate a naive bayes model.
}
