% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_Model.R
\name{test_Model}
\alias{test_Model}
\title{Test prediction model for immunotherapy response}
\usage{
test_Model(Model, SE)
}
\arguments{
\item{Model}{model generated by build_Model() function}

\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}
}
\description{
Generate a naive bayes model.
}
