% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_Model.R
\name{test_SURV_Model}
\alias{test_SURV_Model}
\title{Test Lasso cox prediction model for immunotherapy response}
\usage{
test_SURV_Model(Model, SE, PT_drop)
}
\arguments{
\item{Model}{model generated by build_Model() function}

\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{PT_drop}{If TRUE, only Untreated patient will be use for model training.}
}
\description{
Test a Lasso cox model.
}
