% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_Model.R
\name{test_SVM_model}
\alias{test_SVM_model}
\title{Test prediction model for immunotherapy response}
\usage{
test_SVM_model(Model, SE, PT_drop)
}
\arguments{
\item{Model}{model generated by build_Model() function}

\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{PT_drop}{If TRUE, only Untreated patient will be use for model training.}
}
\description{
Test a naive bayes model.
}
