% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score_biomk.R
\name{tGE8_grading}
\alias{tGE8_grading}
\title{Giving immunotherapy prognosis using tGE8.}
\usage{
tGE8_grading(exp_mtr)
}
\arguments{
\item{exp_mtr}{An expression matrix which rownames are gene symbol and colnames are sample ID.}
}
\description{
The function will return a vector calculated using tGE8.
}
