% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_BE.R
\name{tigeR_Combat}
\alias{tigeR_Combat}
\title{Remove batch effect.}
\usage{
tigeR_Combat(SE)
}
\arguments{
\item{SE}{a SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}
}
\description{
Generate a naive bayes model.
}
