% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMER_SE.R
\name{TIMER_SE}
\alias{TIMER_SE}
\title{TIMER deconvolution}
\usage{
TIMER_SE(SE, type = "SKCM")
}
\arguments{
\item{SE}{a SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}

\item{type}{the cancer type of data.}
}
\description{
use TIMER to predict TME
}
