% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature_utilities.R
\name{weight_mean_signature}
\alias{weight_mean_signature}
\title{Calculate Signature score with weighted mean.}
\usage{
weight_mean_signature(exp_mtr, Signature)
}
\arguments{
\item{exp_mtr}{An expression matrix which rownames are gene symbol and colnames are sample ID.}

\item{Signature}{Gene Signature.}
}
\description{
Calculate Signature score with weighted mean.
}
