% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME_deconvolution.R
\name{xCell}
\alias{xCell}
\title{xCell deconvolution}
\usage{
xCell(SE, ...)
}
\arguments{
\item{SE}{an SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}

\item{...}{the arguments}
}
\description{
use xCell to predict TME
}
