% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature_utilities.R
\name{ZScore_PCA_signature}
\alias{ZScore_PCA_signature}
\title{Calculate Signature score with ZScore and PCA.}
\usage{
ZScore_PCA_signature(exp_mtr, Signature)
}
\arguments{
\item{exp_mtr}{An expression matrix which rownames are gene symbol and colnames are sample ID.}

\item{Signature}{Gene Signature.}
}
\description{
Calculate Signature score with ZScore and PCA.
}
