% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_TME.R
\name{ABIS}
\alias{ABIS}
\title{abis deconvolution}
\usage{
ABIS(SE, sig_matrix)
}
\arguments{
\item{SE}{a SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}

\item{sig_matrix}{gene expression matrix from isolated cells.}
}
\description{
use abis to predict TME
}
