% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_biomk.R
\name{browse_biomk}
\alias{browse_biomk}
\title{Browse which feature in a signature contribute most to the predictive performance}
\usage{
browse_biomk(SE, feature, colors = NULL, PT_drop = FALSE)
}
\arguments{
\item{feature}{dd}

\item{colors}{description}

\item{PT_drop}{description}
}
\description{
generate an integration diagram comprising a bar plot representing AUC and a dot plot denoting Hazard Ratio and P-value.
}
