% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Model.R
\name{build_Model}
\alias{build_Model}
\title{Build machine learning prediction model for immunotherapy response}
\usage{
build_Model(
  SE,
  mtr,
  meta,
  Model,
  feature_genes,
  rmBE = FALSE,
  response_NR = TRUE,
  PT_drop = TRUE,
  ...
)
}
\arguments{
\item{SE}{the dataset you wish to use to build your model. A SummarizedExperiment (SE) object, which can be either a single SE object or a list of SE objects. Note that for each SE object, the colData must contain treatment information under the column name Treatment.}

\item{mtr}{the dataset you wish to use to build your model. A SummarizedExperiment (SE) object, which can be either a single SE object or a list of SE objects. Note that for each SE object, the colData must contain treatment information under the column name Treatment.}

\item{meta}{refers to the specific set of genes you wish to use for model construction.}

\item{Model}{represents the type of model you want to build. You have several options to choose from: "NB" for Naive Bayes, "SVM" for Support Vector Machine, "RF" for Random Forest, "CC" for Cancerclass, "ADB" for Adaboost, "LGB" for Logitboost, and "LGT" for Logistics.}

\item{feature_genes}{refers to the specific set of genes you wish to use for model construction.}

\item{rmBE}{whether remove batch effect between different data set using internal Combat method}

\item{response_NR}{If TRUE, only use R or NR to represent Immunotherapy response of patients.}

\item{PT_drop}{If TRUE, only Untreated patient will be use for model training.}

\item{...}{the arguments}
}
\description{
Generate immunotherapy prognosis prediction model.
}
