% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Model.R
\name{build_NB_model}
\alias{build_NB_model}
\title{Build naive bayes prediction model for immunotherapy response}
\usage{
build_NB_model(
  SE,
  Signature,
  rmBE = FALSE,
  response_NR = TRUE,
  laplace = 1,
  PT_drop,
  ...
)
}
\arguments{
\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{Signature}{an gene set you interested in}

\item{rmBE}{whether remove batch effect between different data set using internal Combat method}

\item{response_NR}{If TRUE, only use R or NR to represent Immunotherapy response of patients.}

\item{laplace}{positive double controlling Laplace smoothing. The default (0) disables Laplace smoothing.}

\item{PT_drop}{If TRUE, only Untreated patient will be use for model training.}

\item{...}{the arguments}
}
\description{
Generate a naive bayes model.
}
