% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Model.R
\name{build_RF_model}
\alias{build_RF_model}
\title{Build random forest prediction model for immunotherapy response}
\usage{
build_RF_model(SE, Signature, rmBE = FALSE, response_NR = TRUE)
}
\arguments{
\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{Signature}{an gene set you interested in}

\item{rmBE}{whether remove batch effect between different data set using internal Combat method}

\item{response_NR}{If TRUE, only use R or NR to represent Immunotherapy response of patients.}
}
\description{
Generate a random forest model.
}
