% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{build_SVM_model}
\alias{build_SVM_model}
\title{perform SVM prediction model.}
\usage{
build_SVM_model(SE, Signature, rmBE = TRUE)
}
\arguments{
\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{Signature}{an gene set you interested in}

\item{rmBE}{whether remove batch effect between different data set using internal Combat method}
}
\description{
Generate a Support Vector Machine model.
}
