% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Signature_Score.R
\name{calculate_Signature_Score}
\alias{calculate_Signature_Score}
\title{Calculating Signature score of existing immunotherapy prognosis Signature.}
\usage{
calculate_Signature_Score(exp_mtr)
}
\arguments{
\item{exp_mtr}{an expression matrix for which you want to calculate the Signature Score.}
}
\description{
The function will return a vector calculated using T cell–inflamed GEP score.
}
