% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_name_standardization.R
\name{cell_name_standardization}
\alias{cell_name_standardization}
\title{Standardize the cell names in TME deconvolution}
\usage{
cell_name_standardization(SE)
}
\arguments{
\item{SE}{a SummarizedExperiment object contains the cell fraction of deconvolution.}
}
\description{
standarize the cell names in TME deconvolution.
}
