% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT.R
\name{CIBERSORT}
\alias{CIBERSORT}
\title{Cibersort functions}
\usage{
CIBERSORT(sig_matrix, mixture_file, perm = 0, QN = TRUE)
}
\arguments{
\item{sig_matrix}{file path to gene expression from isolated cells}

\item{mixture_file}{heterogenous mixed expression}

\item{perm}{Number of permutations}

\item{QN}{Perform quantile normalization or not (TRUE/FALSE)}
}
\description{
Cibersort functions
}
