% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_TME.R
\name{CIBERSORT}
\alias{CIBERSORT}
\title{Cibersort functions}
\usage{
CIBERSORT(
  SE,
  sig_matrix,
  perm = 0,
  QN = TRUE,
  style = "elegant",
  group_color = c("#5f96e8CC", "#ee822fCC")
)
}
\arguments{
\item{SE}{a SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}

\item{sig_matrix}{gene expression matrix from isolated cells.}

\item{perm}{the number of permutations.}

\item{QN}{whether perform quantile normalization or not (TRUE/FALSE).}

\item{style}{the ploting style. c("raw,"elegant","brief")}

\item{group_color}{the color of Responder and Non_Responder.}
}
\description{
Cibersort functions which perform deconvolution to bulk RNA-seq data. And return the a list which first element is cell fraction and second is a box plot.
}
