% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_targetVal}
\alias{compute_targetVal}
\title{Compute target values for ensemble learning architecture}
\usage{
compute_targetVal(
  QC_num,
  sampleType,
  batchID,
  targetVal_method = c("mean", "median"),
  targetVal_batchWise = FALSE,
  targetVal_removeOutlier = !targetVal_batchWise,
  coerce_numeric = FALSE
)
}
\arguments{
\item{QC_num}{a numeric data.frame including the matabolite values of quality control (QC) samples. Row: sample. Column: metabolite variable. see Examples.}

\item{sampleType}{a vector corresponding to \code{QC_num} to specify the type of each sample. see Examples.}

\item{batchID}{a vector corresponding to \code{QC_num} to specify the batch of each sample. see Examples.}

\item{targetVal_method}{a character string specifying how the target values are computed. Can be \code{"mean"} (default) or \code{"median"}.}

\item{targetVal_batchWise}{logical. If \code{TRUE}, the target values will be computed based on each batch, otherwise, based on the whole dataset. Setting \code{TRUE} might be useful if your dataset has very obvious batch effects, but this may also make the algorithm less robust. Default: \code{FALSE}.}

\item{targetVal_removeOutlier}{logical. If \code{TRUE}, outliers will be removed before the computation. Outliers are determined with the 1.5 * IQR (interquartile range) rule. We recommend turning this off when the target values are computed based on batches. Default: \code{!targetVal_batchWise}.}

\item{coerce_numeric}{logical. If \code{TRUE}, values in \code{QC_num} will be coerced to numeric before the computation. Default: \code{FALSE}.}
}
\description{
Compute target values for ensemble learning architecture
}
\details{
Code for checking outliers is adapted from \code{\link[grDevices]{boxplot.stats}}.
}
