% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataloader.R
\name{Dataloader}
\alias{Dataloader}
\title{Process data before running machine learning algorithm}
\usage{
Dataloader(pick = NULL)
}
\arguments{
\item{pick}{a number(1-20) or a numeric vector specify the corresponding dataset(s) you wish to load. Alternatively, you can use Dataloader() with pick=NULL to get an overview of all available datasets.}
}
\description{
Process data before running machine learning algorithm
}
