% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{dataPreprocess}
\alias{dataPreprocess}
\title{Calculate the correlations between user defined gene or gene set and TCGA cancerous samples.}
\usage{
dataPreprocess(exp_mtr, Signature)
}
\description{
Generating an Heatmap showing the pearson correlation coefficient.
}
