% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME_deconvolution.R
\name{deconv_TME}
\alias{deconv_TME}
\title{Cibersort functions}
\usage{
deconv_TME(SE, method, ...)
}
\arguments{
\item{SE}{an SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}

\item{method}{the TME analysis method you want to apply for.}

\item{...}{other argument}
}
\description{
perform Tumor Microenvironment deconvolution through 10 open-source algorithms.
}
