% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{diff_gene}
\alias{diff_gene}
\title{differential gene}
\usage{
diff_gene(SE)
}
\arguments{
\item{SE}{a SummarizedExperiment(SE) object or a list consists of multiple SE objects. The colData of the SE object(s) must contain treatment information named Treatment.}
}
\description{
return differential expression gene between Responder and Non-Responder.
}
