% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_TME.R
\name{diff_TME}
\alias{diff_TME}
\title{Visualize differential analysis cell_fraction (Responder vs NonResponder or Pre-Treatment vs Post-Treatment).}
\usage{
diff_TME(
  SE,
  feature = NULL,
  style = "elegant",
  group_color = c("#5f96e8CC", "#ee822fCC")
)
}
\arguments{
\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{feature}{the feature or features you are interested in.}

\item{style}{description}

\item{group_color}{description}
}
\description{
Visualize differential analysis cell_fraction (Responder vs NonResponder or Pre-Treatment vs Post-Treatment).
}
