% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_TME.R
\name{epic}
\alias{epic}
\title{epic deconvolution}
\usage{
epic(SE, ...)
}
\arguments{
\item{SE}{a SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}

\item{...}{the arguments}
}
\description{
use epic to predict TME
}
