% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_label}
\alias{extract_label}
\title{extract and modify response data for naive bayes model}
\usage{
extract_label(datasetNames)
}
\description{
extract and modify response data for naive bayes model
}
