% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_pie.R
\name{fraction_pie}
\alias{fraction_pie}
\title{Draw pie charts for TME deconvolution result}
\usage{
fraction_pie(mtr, feature, rows = 1)
}
\arguments{
\item{mtr}{cell fraction matrix}

\item{feature}{the color levels of the cell types.}

\item{rows}{the number of rows in the plotting matrix.}
}
\description{
Generate a pie charts illustrating the cell fraction or relative cell abundance for each sample.
}
