% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Gini_gene}
\alias{Gini_gene}
\title{Ranking features in matrix with Gini index}
\usage{
Gini_gene(SE)
}
\arguments{
\item{SE}{a SummarizedExperiment object or a list consists of SE objects. The colData of SE objects must contain response information.}
}
\description{
By calculating the Gini index of different genes, you can get an overview of the classification efficiency of different genes.
}
