% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gini_gene}
\alias{gini_gene}
\title{Ranking features in matrix with Gini index}
\usage{
gini_gene(SE)
}
\arguments{
\item{SE}{a SummarizedExperiment object or a list consists of SE objects. The colData of SE objects must contain response information.}
}
\description{
calculating the Gini index and get an overview of the classification efficiency of genes.
}
