% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Gini_rank}
\alias{Gini_rank}
\title{Ranking features in matrix with Gini index}
\usage{
Gini_rank(SE, ascending = TRUE)
}
\arguments{
\item{SE}{a SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{ascending}{If ascending = TRUE, the result will be display in ascending order.}
}
\description{
By calculating the Gini index of different genes, you can get an overview of the classification efficiency of different genes.
}
