% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunotherapy_Response.R
\name{Immunotherapy_Response}
\alias{Immunotherapy_Response}
\title{Perform differential expression analysis and survival analysis.}
\usage{
Immunotherapy_Response(gene, SE)
}
\arguments{
\item{gene}{The gene which you wanted.}

\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain treatment information names Treatment.}
}
\description{
Perform differential expression analysis and survival analysis in certain gene and return the result.
}
