% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunotherapy_Response.R
\name{Immunotherapy_Response}
\alias{Immunotherapy_Response}
\title{Perform differential expression analysis and survival analysis.}
\usage{
Immunotherapy_Response(SE, geneSet = NULL, method = NULL, PT_drop = TRUE)
}
\arguments{
\item{SE}{a SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain treatment information names Treatment.}

\item{geneSet}{The geneSet which you wanted.}

\item{method}{the method for calculating gene set scores which has several options: Average_mean, Weighted_mean, or GSVA. The method should be set to NULL if the length of the parameter geneSet is 1. This means that if you are working with only one gene, the specific calculation method may not be applicable or necessary.}

\item{PT_drop}{If TRUE, only Untreated patient will be use for model training.}
}
\description{
Perform differential expression analysis and survival analysis in certain gene and return the result.
}
