% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME_deconvolution.R
\name{IPS}
\alias{IPS}
\title{IPS deconvolution}
\usage{
IPS(SE, project = NULL, plot = FALSE)
}
\arguments{
\item{SE}{an SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}

\item{project}{project}

\item{plot}{if TRUE return the plot}
}
\description{
use xCell to predict TME
}
