% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matrix_t.test}
\alias{matrix_t.test}
\title{differential gene}
\usage{
matrix_t.test(V, P, N)
}
\arguments{
\item{V}{the expression vector of a gene}

\item{P}{the index of Positive samples in vector V}

\item{N}{the index of Negative samples in vector V}
}
\description{
Return differential expression gene between Responder and Non-Responder
}
