% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plt_diff}
\alias{plt_diff}
\title{plot differential result(Responder vs Non-Responder or Pre-Treatment vs Post-Treatment)}
\usage{
plt_diff(SE, gene, type, method = "Average_mean")
}
\arguments{
\item{SE}{SE an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{gene}{is the Gene or Gene set you are interested in.}

\item{type}{'Treatment' or 'Response'.the type of analysis you want to perform(Responder vs Non-Responder or Pre-Treatment vs Post-Treatment)}

\item{method}{the method for calculating gene set scores which has several options: Average_mean, Weighted_mean, or GSVA. The method can be set to NULL if the length of the parameter geneSet is 1. This means that if you are working with only one gene, the specific calculation method may not be applicable or necessary.}
}
\description{
plot differential result(Responder vs Non-Responder or Pre-Treatment vs Post-Treatment)
}
