% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plt_Preprocess}
\alias{plt_Preprocess}
\title{Prepare data for plot}
\usage{
plt_Preprocess(gene, SE, method, type, PT_drop, log_sc)
}
\arguments{
\item{gene}{is the Gene or Gene set you are interested in.}

\item{SE}{a SummarizedExperiment(SE) object or a list consists of multiple SE objects. The colData of the SE object(s) must contain treatment information named Treatment.}

\item{method}{the method for calculating gene set scores. Can be NULL if the length of parameter gene is 1.}

\item{type}{the type of information}

\item{PT_drop}{If TRUE, only Untreated patient will be use for model training.}

\item{log_sc}{if TRUE, log(value + 1)}
}
\description{
Preparing data for ploting.
}
