% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plt_style}
\alias{plt_style}
\title{Build plot theme}
\usage{
plt_style(df, textcol)
}
\arguments{
\item{df}{a dataframe}

\item{textcol}{the color of the text in the plot}
}
\description{
return the ploting theme
}
