% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plt_surv}
\alias{plt_surv}
\title{plot differential result(Pre-Treatment vs Post-Treatment)}
\usage{
plt_surv(SE, gene, method = "Average_mean")
}
\arguments{
\item{SE}{SE an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{gene}{is the Gene or Gene set you are interested in.}

\item{method}{the method for calculating gene set scores. Can be NULL if the length of parameter gene is 1.}
}
\description{
The association between gene expression and overall survival in the immunotherapy data was calculated using univariate Cox regression analysis.
}
