% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plt_surv}
\alias{plt_surv}
\title{plot differential result(Pre-Treatment vs Post-Treatment)}
\usage{
plt_surv(
  SE,
  gene,
  method = "Average_mean",
  style = "elegant",
  conf.int = FALSE,
  val.pos = c(0.03, 0.2),
  lg.pos = c(0.6, 0.9)
)
}
\arguments{
\item{SE}{SE an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information.}

\item{gene}{is the Gene or Gene set you are interested in.}

\item{method}{the method for calculating gene set scores which has several options: Average_mean, Weighted_mean, or GSVA. The method can be set to NULL if the length of the parameter geneSet is 1. This means that if you are working with only one gene, the specific calculation method may not be applicable or necessary.}

\item{style}{the ploting style. c("raw,"elegant","brief")}

\item{conf.int}{logical value. If TRUE, plots confidence interval.}

\item{val.pos}{the position of annotation value.}

\item{lg.pos}{the position of legend. When lg.pos=c(0,0), the legend will be placed at the leftdown of the plot.}
}
\description{
The association between gene expression and overall survival in the immunotherapy data was calculated using univariate Cox regression analysis.
}
