% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_biomk.R
\name{roc_biomk}
\alias{roc_biomk}
\title{Assessing the impact of Signature on Immunotherapy Response}
\usage{
roc_biomk(
  SE,
  Signature,
  method = NULL,
  rmBE = FALSE,
  response_NR = TRUE,
  PT_drop = TRUE,
  auc.pos = c(0.3, 0.42),
  auc.round = 3,
  textcol = "#646464"
)
}
\arguments{
\item{SE}{the dataset you wish to use to test your Signature. A SummarizedExperiment (SE) object, which can be either a single SE object or a list of SE objects. Note that for each SE object, the colData must contain treatment information under the column name Treatment.}

\item{Signature}{The gene which you wanted.}

\item{method}{the method for calculating gene set scores which has several options: Average_mean, Weighted_mean, or GSVA. The method can be set to NULL if the length of the parameter geneSet is 1. This means that if you are working with only one gene, the specific calculation method may not be applicable or necessary.}

\item{rmBE}{whether remove batch effect between different data set using internal Combat method}

\item{response_NR}{If TRUE, classify patients with CR, MR, PR as Responders (R), and those with PD, SD, NR as Non-Responders(NR).}

\item{PT_drop}{If TRUE, only Untreated patient will be use for model training.}

\item{auc.pos}{the position of the AUC value}

\item{auc.round}{the decimal places you want to keep for auc value}

\item{textcol}{the color of the text in the plot}
}
\description{
generate a Receiver Operating Characteristic (ROC) object and a curve to assess the predictive performance.
}
