% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{select_variable}
\alias{select_variable}
\title{Select variables for ensemble learning architecture}
\usage{
select_variable(
  train_num,
  test_num = NULL,
  train_batchID = NULL,
  test_batchID = NULL,
  correlation_batchWise = FALSE,
  correlation_type = c("cor", "pcor"),
  correlation_method = c("spearman", "pearson"),
  min_var_num = 5,
  max_var_num = 10,
  coerce_numeric = FALSE
)
}
\arguments{
\item{train_num}{a numeric data.frame including the matabolite values of training samples (can be quality control samples). Row: sample. Column: metabolite variable. see Examples.}

\item{test_num}{a optional numeric data.frame including the matabolite values of test samples (can be subject samples). Row: sample. Column: metabolite variable. see Examples. If \code{NULL}, the variables will be selected based on \code{train_num} only.}

\item{train_batchID}{ignored if \code{correlation_batchWise = FALSE}. \code{NULL} or a vector corresponding to \code{train_num} to specify the batch of each sample. see Examples.}

\item{test_batchID}{ignored if \code{correlation_batchWise = FALSE}. \code{NULL} or a vector corresponding to \code{test_num} to specify the batch of each sample. see Examples.}

\item{correlation_batchWise}{logical. Specify whether the variable selection should be performed for each batch. Setting \code{TRUE} might be useful if your dataset has very obvious batch effects, but this may also make the algorithm less robust. Default: \code{FALSE}.}

\item{correlation_type}{a character string indicating correlation (\code{"cor"}, default) or partial correlation (\code{"pcor"}) is to be used. Can be abbreviated. Note: computing partial correlations of a large dataset can be very time-consuming.}

\item{correlation_method}{a character string indicating which correlation coefficient is to be computed. One of \code{"spearman"} (default) or \code{"pearson"}. Can be abbreviated.}

\item{min_var_num}{an integer specifying the minimum number of the selected variables. If \code{NULL}, no limited, but 1 at least. Default: 5.}

\item{max_var_num}{an integer specifying the maximum number of the selected variables. If \code{NULL}, no limited, but \code{ncol(train_num)} at most. Default: 10.}

\item{coerce_numeric}{logical. If \code{TRUE}, values in \code{train_num} and  \code{test_num} will be coerced to numeric before the computation. Default: \code{FALSE}.}
}
\description{
Select variables for ensemble learning architecture
}
