% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{surv_styling}
\alias{surv_styling}
\title{the style of survival plot}
\usage{
surv_styling(df, style, conf.int, gene, method, val.pos, lg.pos)
}
\arguments{
\item{df}{the data source}

\item{style}{plotting style}

\item{conf.int}{logical value. If TRUE, plots confidence interval.}

\item{gene}{the gene you interested in.}

\item{method}{the method}

\item{val.pos}{the position of annotation value.}

\item{lg.pos}{the position of legend.}
}
\description{
the style of survival plot
}
