% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunotherapy_Response.R
\name{survival_Score}
\alias{survival_Score}
\title{Calculating survival score of patients.}
\usage{
survival_Score(gene = "CD274", SE)
}
\arguments{
\item{gene}{The gene which you wanted.}

\item{SE}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain event time(names time), event(names status).}
}
\description{
Survival score was calculated using the following formula: âˆ’ğ‘†ğ¼ğºğ‘(ğ‘™ğ‘œğ‘”2(ğ»ğ‘…)) Ã— ğ‘™ğ‘œğ‘”10(ğ‘ƒ), where ğ»ğ‘… represents the hazard ratio and ğ‘ƒ represents the P value derived from univariate Cox regression analysis.
}
