% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunotherapy_Response.R
\name{survival_Score}
\alias{survival_Score}
\title{Calculating survival score of Untreated patients.}
\usage{
survival_Score(Score, meta, PT_drop)
}
\arguments{
\item{Score}{an SummarizedExperiment(SE) object or a list consists of SE objects. The colData of SE objects must contain response information names response.}

\item{meta}{The geneSet which you wanted.}

\item{PT_drop}{If TRUE, only Untreated patient will be use for model training.}
}
\description{
Survival score was calculated using the following formula: âˆ’ğ‘†ğ¼ğºğ‘(ğ‘™ğ‘œğ‘”2(ğ»ğ‘…)) Ã— ğ‘™ğ‘œğ‘”10(ğ‘ƒ), where ğ»ğ‘… represents the hazard ratio and ğ‘ƒ represents the P value derived from univariate Cox regression analysis.
}
