% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME_deconvolution.R
\name{TIMER}
\alias{TIMER}
\title{TIMER deconvolution}
\usage{
TIMER(SE, type = "SKCM")
}
\arguments{
\item{SE}{an SummarizedExperiment object contains the bulk RNA-seq dataset that you want to use for deconvolution and obtaining its cell fraction.}

\item{type}{the cancer type of data.}
}
\description{
use TIMER to predict TME
}
