% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME_deconvolution.R
\name{TIMER}
\alias{TIMER}
\title{TIMER deconvolution}
\usage{
TIMER(exp_mtr, type)
}
\arguments{
\item{exp_mtr}{matrix or data.frame with features in rows and samples in columns}

\item{type}{type of cancer}
}
\description{
use TIMER to predict TME
}
