% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME_deconvolution.R
\name{TIMER}
\alias{TIMER}
\title{TIMER deconvolution}
\usage{
TIMER(exp_mtr, type)
}
\arguments{
\item{exp_mtr}{a matrix or data.frame with genes in rows and patients in columns.}

\item{type}{the cancer type of data.}
}
\description{
use TIMER to predict TME
}
